cashed_get modified

+++ spec/web_helper/web_helper_spec.rb
@@ -1,23 +1,40 @@
 require_relative '../../lib/web_helper'
 require 'rspec'
+require 'webmock/rspec'
 
 describe WebHelper do
+  RSpec::Matchers.define_negated_matcher :not_have_requested, :have_requested
+
 	let(:url) { "http://www.imdb.com/title/tt0111161/?pf_rd_m=A2FGELUUNOQJNL&pf_rd_p=2398042102&pf_rd_r=1WHNSFSC9ND0X52YFRAT&pf_rd_s=center-1&pf_rd_t=15506&pf_rd_i=top&ref_=chttp_tt_1" }
   let(:url2) { "http://www.imdb.com/title/tt0111161?pf_rd_m=A2FGELUUNOQJNL&" }
+  let(:file_name) { WebHelper.get_name_for(url) }
+  let(:file_contents) { File.read(file_name) }
 
   describe '#parse_url' do
     it { expect(WebHelper.parse_url(url)).to eq('title_tt0111161.html') }
     it { expect(WebHelper.parse_url(url2)).to eq('title_tt0111161.html') }
   end
 
-  describe '#file_create' do
-    let(:file_name) { WebHelper.file_create(url) }
-
-    after(:example) do
-      File.delete(file_name)
+  describe '#cashed_get' do
+    before(:example) do
+      stub_request(:any, url)
     end
 
-    it { expect(File.exists?(file_name)).to be true }
+    context 'when no file' do
+      before(:example) do
+        File.delete(file_name) if File.exists?(file_name)
+      end
+
+      it { expect(WebHelper.cashed_get(url)).to eq(file_contents).and have_requested(:any, url) }
+    end
+
+    context 'when file exists' do
+      after(:example) do
+        File.delete(file_name) if File.exists?(file_name)
+      end
+
+      it { expect(WebHelper.cashed_get(url)).to not_have_requested(:any, url).and eq(file_contents) }
+    end
   end
 
 end
